move Points class into classes.py
