+++ plantcv/plantcv/classes.py
@@ -1,7 +1,11 @@
 # PlantCV classes
 import os
+import cv2
 import json
 from plantcv.plantcv import fatal_error
+import matplotlib.pyplot as plt
+from math import floor
+
 
 
 class Params:
@@ -242,3 +246,39 @@ class PSII_data:
             protocol: xr.DataArray with name equivalent to initialized attributes
         """
         self.__dict__[protocol.name] = protocol
+
+
+class Points(object):
+    """Point annotation/collection class
+    """
+
+    def __init__(self, img, figsize=(12, 6)):
+        """
+        Initialization
+        :param img: image data
+        :param figsize: desired figure size, (12,6) by default
+        """
+
+        self.fig, self.ax = plt.subplots(1, 1, figsize=figsize)
+        self.ax.imshow(cv2.cvtColor(img, cv2.COLOR_BGR2RGB))
+
+        self.points = []
+        self.events = []
+
+        self.fig.canvas.mpl_connect('button_press_event', self.onclick)
+
+    def onclick(self, event):
+        self.events.append(event)
+        if event.button == 1:
+
+            self.ax.plot(event.xdata, event.ydata, 'x', c='red')
+            self.points.append((floor(event.xdata), floor(event.ydata)))
+
+        else:
+            idx_remove, _ = _find_closest_pt((event.xdata, event.ydata), self.points)
+            # remove the closest point to the user right clicked one
+            self.points.pop(idx_remove)
+            ax0plots = self.ax.lines
+            self.ax.lines.remove(ax0plots[idx_remove])
+        self.fig.canvas.draw()
+
